Move generic observe events to ParallelAlgorithms
