+++ tests/Unit/ParallelAlgorithms/Events/CMakeLists.txt
@@ -1,0 +1,21 @@
+# Distributed under the MIT License.
+# See LICENSE.txt for details.
+
+set(LIBRARY "Test_ParallelAlgorithmsEvents")
+
+set(LIBRARY_SOURCES
+  Test_ObserveErrorNorms.cpp
+  Test_ObserveFields.cpp
+  )
+
+add_test_library(
+  ${LIBRARY}
+  "ParallelAlgorithms/Events/"
+  "${LIBRARY_SOURCES}"
+  "DataStructures;Domain;ErrorHandling;Time;Utilities"
+  )
+
+add_dependencies(
+  ${LIBRARY}
+  module_ConstGlobalCache
+  )
